requirements.txt
async-timeout==4.0.2
bcrypt==3.2.2
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==4.4.2
Flask-Limiter==2.5.0
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
limits==2.7.0
MarkupSafe==2.1.1
packaging==21.3
pep8==1.7.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
rich==12.5.1
six==1.16.0
SQLAlchemy==1.4.39
svgwrite==1.4.3
typing_extensions==4.3.0
wavedrom==2.0.3.post3.dev16+g05b1220
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
